Version 1
Decided to create a jQuery History project that dynamically retrieves
content and updates it on the page.

The history part is all about the fact that the browser's history is
manipulated so you can go back and forward in the historic timeline to
another dynamically loaded page, and the URL in the address bar also
updates, so those who wish to copy the URL and share it will direct
traffic to the right content.

It was a fun little project to create, and I hope it helps someone out
there make something awesome. I may update it in the future, but at this
point, I consider this to be my initial and final commit of a fully
working project. Enjoy. :)