Added DirectoryIndex rule to .htaccess file
When testing the code before this update in a local environment,
everything worked fine, but when I uploaded it to my server, it didn't
work as expected. This is due to another .htaccess file I have in my
site's root directory. Simply adding one line of code to this file fixed
the issue.