Removed URL prefix from AJAX request in custom.js
One of the problems with working in a local testing enviroment is the
hostname is usually different to that of the live site. At one point in
my code I tested to see if adding this prefix would fix an issue, and
realise now that I didn't need it at all, which makes things work both
locally and on the live site.